+++ packages/koot-boilerplate/src/components/center/index.d.ts
@@ -1,0 +1,4 @@
+import { Component } from 'react';
+
+const Center = Component;
+export default Center;