+++ packages/koot/globa.d.ts
@@ -1,1 +1,0 @@
-declare function __(...translateKeys: Array<string>): String;